# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Logueo",
        "description": "Endpoints para gestión de logueo",
        "contact": {
            "email": "email@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/familias": {
            "get": {
                "tags": [
                    "Familias"
                ],
                "summary": "Obtener todos los familias",
                "description": "Obtener todas las familias.",
                "operationId": "edd5027b516710936f866b84deaaf72c",
                "responses": {
                    "200": {
                        "description": "Lista de familias",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "familias": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Familia"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Familias"
                ],
                "summary": "Crear un nuevo familia",
                "description": "Crear una nueva familia.",
                "operationId": "458cbb1ab4e0a6eb58ad576ed5a048fe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Familia"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Familia creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Familia creado correctamente"
                                        },
                                        "familia": {
                                            "$ref": "#/components/schemas/Familia"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la validación de datos"
                    }
                }
            }
        },
        "/api/familias/{familia}": {
            "put": {
                "tags": [
                    "Familias"
                ],
                "summary": "Actualizar un familiass existente",
                "description": "Actualizar una familia existente.",
                "operationId": "88b40b64910ee6714b361e97be366216",
                "parameters": [
                    {
                        "name": "familia",
                        "in": "path",
                        "description": "ID del familia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Familia"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Familia actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Familia actualizado correctamente"
                                        },
                                        "familia": {
                                            "$ref": "#/components/schemas/Familia"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la validación de datos"
                    },
                    "404": {
                        "description": "Familia no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Familias"
                ],
                "summary": "Eliminar un familia",
                "description": "Eliminar una familia.",
                "operationId": "8b777a325c4ef2117a8d4345afc3b28e",
                "parameters": [
                    {
                        "name": "familia",
                        "in": "path",
                        "description": "ID del familia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Familia eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Familia eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Familia no encontrado"
                    }
                }
            }
        },
...
        "/api/inscripciones/{inscripcion}": {
            "put": {
                "tags": [
                    "Inscripciones"
                ],
                "summary": "Actualizar una inscripcion existente",
                "description": "Actualizar una inscripcion existente.",
                "operationId": "5b773971dbfb51da5efc6e49ecce5ce8",
                "parameters": [
                    {
                        "name": "inscripcion",
                        "in": "path",
                        "description": "ID del inscripcion",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Inscripcion"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inscripcion actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Inscripcion actualizada correctamente"
                                        },
                                        "inscripcion": {
                                            "$ref": "#/components/schemas/Inscripcion"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la validación de datos"
                    },
                    "404": {
                        "description": "Inscripcion no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Familia": {
                "title": "Familia",
                "required": [
                    "nombre"
                ],
                "properties": {
                    "id": {
                        "description": "ID único de la familia",
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "description": "Nombre de la familia",
                        "type": "string",
                        "example": "CFGS Estética Integral y Bienestar"
                    }
                },
                "type": "object"
            },
            "Inscripcion": {
                "title": "Inscripcion",
                "required": [
                    "comentarios",
                    "equipo_id",
                    "estado_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID único de la inscripcion",
                        "type": "integer",
                        "example": 1
                    },
                    "comentarios": {
                        "description": "Contenido adicional de la inscripcion",
                        "type": "string",
                        "example": "Componentes con gran posibilidad de aceptacion"
                    },
                    "equipo_id": {
                        "description": "ID del equipo asociado a la inscripcion",
                        "type": "integer",
                        "example": "3"
                    },
                    "estado_id": {
                        "description": "ID del estado de la inscripcion",
                        "type": "integer",
                        "example": "3"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}